requirements.txt
fastapi
uvicorn
confluent-kafka
pydantic-settings
watchdog
sqlalchemy
asyncpg
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
faker
pymilvus